Move OutdatedPullRequestStrategy.ts to core dir
