+++ src/core/index.ts
@@ -3,6 +3,7 @@ export * from './OutdatedPackage'
 export * from './OutdatedPackageProcessor'
 export * from './OutdatedPackageProcessorCreator'
 export * from './OutdatedPackagesProcessor'
+export * from './OutdatedPullRequestStrategy'
 export * from './PackageUpdater'
 export * from './Recreate'
 export * from './SucceededResult'
