+++ src/github/pull-request/creator/index.ts
@@ -1,6 +1,5 @@
 export * from './AssigneesAdder'
 export * from './GitHubUrlOptimizer'
-export * from './OutdatedPullRequestStrategy'
 export * from './PackageDiffsSectionCreator'
 export * from './PullRequestBodyCreator'
 export * from './PullRequestCreator'
