+++ src/options/Options.test.ts
@@ -1,4 +1,4 @@
-import { OutdatedPullRequestStrategy } from '../github'
+import { OutdatedPullRequestStrategy } from '../core'
 import { LogLevel } from '../logger'
 import { PackageManagerName } from '../package-manager'
 import {
